Add THSupport to allow overriding of setupBuildFlags
This is fairly nasty.  We rely on overriding  for TH support, but this then prevents us
from allowing to override  in final flake.nix or similar expressions for the final build
package/component/executable.  This even fails silently and TH just doesn't work.

This seems to be a larger topic around how we deal with TH, and module inheritance. We should not override
existing keys, but namespace them properly, and thus still allow overriding.